% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sealevel.functions.r
\name{find.stations.f}
\alias{find.stations.f}
\title{Find variables with a keyword or phrase}
\usage{
find.stations.f(search.term, plot = F)
}
\arguments{
\item{search.term}{a term to search, e.g. "halifax", "tokyo"}

\item{plot}{if true it will plot the locations of all the tide stations and the ones matching the search criteria in red}
}
\description{
Does a fuzzy search for the term in the station list and returns all the information from rows in the station table
              that contains the term. Case insensitive.
}
\examples{
find.stations.f("halifax")
       find.stations.f("MumBaI",T)
}
\author{
Daniel Duplisea
}
