% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sealevel.functions.r
\name{limits}
\alias{limits}
\title{Outer decimal bounding limits for plotting}
\usage{
limits(vector, places = 100)
}
\arguments{
\item{vector}{a vector of numbers to find the limits}

\item{places}{number of digits. you need to set the level of rounding by setting the places by factors of 10. So if 100 it will take the bottom limit to the lowest value in the series to closest 100 value of the lower side and vice versa}
}
\description{
Finds a bounding limit for plotting a vector that is rounded outside the data range
}
\author{
Daniel Duplisea
}
