% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sealevel.functions.r
\name{sealevel.f}
\alias{sealevel.f}
\title{Get and plot sea level data and determine rate of change in level from PSMSL database}
\usage{
sealevel.f(id, y1 = 1920, y2 = 2000)
}
\arguments{
\item{id}{The tide station id number from psmsl}

\item{y1}{the first year of data you want to include in the plot}

\item{y2}{the breakpoint year you want to force to calculate a new sea level change rate from}
}
\description{
Queries the PSMSL database (psmsl.org) for monthly RLR data from one of its stations that you specify with id. This
              will also fit a linear regression over all the data and calculate the rate of change in sea level (mm/y) from y1 or
              or the earliest data year available if y1 is before that year. It will also plot and calculate the rate of change in
              sea level from y2 until the end of the time series. Finally, it will make a third plot where a segmented regression
              is fitted to the data from y1 until the end of the data series and find the most appropriate single breakpoint
              calculating rate of change in sea level before and after the breakpoint. The segmented regression is fitted using the
              package segmented.
}
\author{
Daniel Duplisea
}
